+++ include/m2/network/TcpSocket.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "IpAddressAndPort.h"
 #include "../Meta.h"
+#include <cstdio>
 
 namespace m2::network {
 	// Forward declarations
@@ -46,8 +47,8 @@ namespace m2::network {
 		/// was non-blocking, but the write operation would have blocked, most likely because the kernel's output buffer
 		/// is full. If the socket is blocking, the call might block until at least some bytes are sent. If unexpected
 		/// is returned, system error message is returned, and the socket shouldn't be used for further communication.
-		expected<ssize_t> send(const uint8_t* buffer, size_t length);
-		expected<ssize_t> send(const char* buffer, size_t length) { return send(reinterpret_cast<const uint8_t*>(buffer), length); }
+		expected<int> send(const uint8_t* buffer, size_t length);
+		expected<int> send(const char* buffer, size_t length) { return send(reinterpret_cast<const uint8_t*>(buffer), length); }
 
 		/// Returns the number of bytes actually received. If 0 is returned, EAGAIN/EWOULDBLOCK has occurred. If the
 		/// socket was non-blocking, EAGAIN means the receive operation would have blocked because there is no data
@@ -55,8 +56,8 @@ namespace m2::network {
 		/// (regardless of blocking or non-blocking) and it has expired. If receive timeout is not set, and the socket
 		/// is blocking, the call would simply block until at least some data is received. If unexpected is returned,
 		/// the system error string is returned. In this case, socket should be closed as no further data is expected.
-		expected<ssize_t> recv(uint8_t* buffer, size_t length);
-		expected<ssize_t> recv(char* buffer, size_t length) { return recv(reinterpret_cast<uint8_t*>(buffer), length); }
+		expected<int> recv(uint8_t* buffer, size_t length);
+		expected<int> recv(char* buffer, size_t length) { return recv(reinterpret_cast<uint8_t*>(buffer), length); }
 
 		// Give Select class access to internals
 		friend Select;
